接続ページを別ファイルへ
